Update patch set 2
Patch Set 2: Code-Review-1

This requires an extra round-trip to voltha, why not just allow the request to go through, and allow voltha to fail the request server-side?
This seems like extra, unnecessary verification that should already be done on the server.

I could see some of this being useful for autocomplete, however.

Patch-set: 2