Update patch set 2
Patch Set 2:

> This requires an extra round-trip to voltha, why not just allow the
 > request to go through, and allow voltha to fail the request
 > server-side?
 > This seems like extra, unnecessary verification that should already
 > be done on the server.
 > 
 > I could see some of this being useful for autocomplete, however.

log is configured via etcd, so there isn't validation on the server side when a log level is sdet.

 > This requires an extra round-trip to voltha, why not just allow the
 > request to go through, and allow voltha to fail the request
 > server-side?
 > This seems like extra, unnecessary verification that should already
 > be done on the server.
 > 
 > I could see some of this being useful for autocomplete, however.

The log is configured by updating keys in etcd, so there isn't a server-side validation or an API call that could fail if there is a typo in the component name.

Patch-set: 2