Update patch set 2
Patch Set 2:

> > The log is configured by updating keys in etcd, so there isn't a
 > > server-side validation or an API call that could fail if there is
 > a
 > > typo in the component name.
 > 
 > What about in a pre-configuration scenario?  In such a case, voltha
 > may not be available, and there may not be any config in etcd.
 > 
 > Perhaps print a warning, but I'm not yet convinced that the request
 > should be rejected at this stage.

the plan, afaik, was to do just that. i.e. if a component or package is not matched then still process the request but give a warning.

i could argue that if a match cannot be found then reject unless there is a `--force` flag as well.

Patch-set: 2