Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Not sure this is the best approach (i.e. just enlarging the max size of the message).  The fix for this is to allow filtering and limiting the number of devices returned on the server

Patch-set: 1
Reviewer: Gerrit User 1000003 <1000003@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=-1