Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (1 comment)
> 
> Not sure this is the best approach (i.e. just enlarging the max size of the message).  The fix for this is to allow filtering and limiting the number of devices returned on the server

It's a quick solution for an existing bug. Having 1000 devices in voltha makes voltctl non-functional, so here is the fix. Introducing pagination is a big change which requires much more work than this fix, so I would propose to create a followup story, so it can be addressed later. As for versioning the changes are pending since 3 weeks, so every 3 weeks it will require a version ;-)

Patch-set: 1
Label: Code-Review=0