Update patch set 2
Patch Set 2:

> Should refer to the latest protos ...

This actually opens an interesting discussion:  Should components be updated to use the latest protos "as soon as possible", or only as needed?  I agree that it's best to update sooner than later, but there's potential for a lot of churn (many commits to many repos for even a small change to the protos).

Patch-set: 2
Reviewer: Gerrit User 1000107 <1000107@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>